A hackish use of the pypvs that just inserts data into the dlcgi request
The pypvs is initialized along the whole sunpower integration. Then on
the fetch evert that already exists, a call to fetch the uptime via
varserver is made. This value is appended to the json feed received
by the dlcgi request. Data is then visualized via the field:
PV Supervisor PVS6 ZT190885000549A1562 Varserver Uptime